Added a new configuration option 'run_handler_in_init'
By default when in html5history mode the route handler is executed in Router.init() because of need for routing cannot be determined (like it easily can when using hashes). Usually this is the wanted behavior, but since that is not always the case this option makes it possible to disable execution.