Make work with Browserify.
This depends on hij1nx/codesurgeon#13 being accepted. Once it is, and the codesurgeon dependency is updated, the existing build script will produce a file that can be used either with <script> tag includes, or with browserify (and any other browser-based CommonJS module loaders).

Furthermore, this adds a "browserify" entry in package.json pointing to the browser-built file, so that users can just do `require("director")` instead of `require("director/build/director-1.1.3")`.